backend/requirements.txt
asgiref==3.8.1
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.1
djangoviz
drf-social-oauth2==3.1.0
dynaconf==3.2.6
factory-boy==3.3.1
flower==2.0.1
googlemaps==4.10.0
gunicorn==21.2.0
hvac==2.3.0
idna==3.8
jwcrypto==1.5.6
loguru==0.7.2
mysqlclient>=2.0
mysql-connector-python
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest-django==4.8.0
pytest==8.3.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn[standard]==0.25.0
validators==0.34.0
watchfiles
whitenoise==6.7.0
win32-setctime==1.1.0